chore: move UploadView to top level
